Factor out damageable
