+++ Assets/Scenes/MainScene.unity
@@ -132,7 +132,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 13784882}
-  m_Layer: 0
+  m_Layer: 8
   m_Name: AttackPointR
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -162,7 +162,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 348015763}
-  m_Layer: 0
+  m_Layer: 9
   m_Name: AttackPointL
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -280,7 +280,7 @@ GameObject:
   - component: {fileID: 596439564}
   - component: {fileID: 596439563}
   - component: {fileID: 596439568}
-  m_Layer: 0
+  m_Layer: 9
   m_Name: Knight
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -339,7 +339,7 @@ Animator:
   m_Avatar: {fileID: 0}
   m_Controller: {fileID: 9100000, guid: 45d66aa1cc92c46fe819a10442b31a74, type: 2}
   m_CullingMode: 0
-  m_UpdateMode: 1
+  m_UpdateMode: 0
   m_ApplyRootMotion: 1
   m_LinearVelocityBlending: 0
   m_WarningMessage: 
@@ -452,7 +452,7 @@ GameObject:
   - component: {fileID: 1363332664}
   - component: {fileID: 1363332666}
   - component: {fileID: 1363332667}
-  m_Layer: 0
+  m_Layer: 8
   m_Name: Player
   m_TagString: Player
   m_Icon: {fileID: 0}
