+++ Assets/Scripts/Damageable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 01c1fcc3c3b4d41fab496119da957698
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
