+++ Assets/Scripts/GoblinController.cs
@@ -5,15 +5,6 @@ using UnityEngine;
 public class GoblinController : MonoBehaviour
 {
     public float speed = 3.0f;
-    public int maxHealth = 3;
-
-    public int health { get { return currentHealth; } }
-    int currentHealth;
-
-    public float timeInvincible = 2.0f;
-    bool isInvincible;
-    float invincibleTimer;
-
 
     Rigidbody2D rigidbody2d;
     float horizontal;
@@ -26,7 +17,6 @@ public class GoblinController : MonoBehaviour
     {
         animator = GetComponent<Animator>();
         rigidbody2d = GetComponent<Rigidbody2D>();
-        currentHealth = maxHealth;
     }
 
     // Update is called once per frame
@@ -47,21 +37,9 @@ public class GoblinController : MonoBehaviour
         //animator.SetFloat("Look X", lookDirection.y);
         animator.SetFloat("Speed", move.magnitude);
 
-        if (isInvincible)
-        {
-            invincibleTimer -= Time.deltaTime;
-            if (invincibleTimer < 0)
-                isInvincible = false;
-        }
-
         if (Input.GetKeyDown(KeyCode.C))
         {
             Attack();
-        }
-
-        if (Input.GetKeyDown(KeyCode.V))
-        {
-            ChangeHealth(-1);
         }
     }
 
@@ -77,31 +55,5 @@ public class GoblinController : MonoBehaviour
     void Attack()
     {
         animator.SetTrigger("Stab");
-    }
-
-    void Die()
-    {
-        Debug.Log("I'm dead!");
-        animator.SetTrigger("Death");
-    }
-
-    public void ChangeHealth(int amount)
-    {
-        if (amount < 0)
-        {
-            if (isInvincible)
-                return;
-
-            animator.SetTrigger("Hit");
-            isInvincible = true;
-            invincibleTimer = timeInvincible;
-        }
-
-        currentHealth = Mathf.Clamp(currentHealth + amount, 0, maxHealth);
-        Debug.Log(currentHealth + "/" + maxHealth);
-        if (currentHealth == 0)
-        {
-            Die();
-        }
     }
 }
